etl/requirements.txt
apache-airflow
psycopg2-binary
elasticsearch
requests
python-dotenv